+++ src/opf_python/jeremy.py
@@ -4,7 +4,7 @@
 import numpy as np
 from universal import find_srBs
 
-def spGrids(lattice,kpd):
+def spGrids(A,kpd):
     """Finds the symmetry preserving grids for a given lattice at a given density.
 
     Args:
@@ -16,24 +16,25 @@ def spGrids(lattice,kpd):
             and packing fraction.
     """
 
-    Bs = find_srBs(A,target,exact=True)
+    Bs = find_srBs(A,kpd,exact=True)
 
     #now we want to find the grid vectors
-    grids = []
+    B_grids = []
     ns = []
     for B in Bs:
-        grids.append(np.transpose(np.linalg.inv(B)))
-        ns.apppend(np.linalg.det(B)/np.linalg.det(A))
+        B_grids.append(np.transpose(np.linalg.inv(B)))
+        ns.append(np.linalg.det(B)/np.linalg.det(A))
     
         #now I want to find the packing fraction and the r_min for each of these grids.
         from phenum.vector_utils import _minkowski_reduce_basis
 
         grids = []
-        for grid in grids:
-            min_grid = _minkowski_reduce_basis(grid)
+        for grid in B_grids:
+            min_grid = np.transpose(_minkowski_reduce_basis(np.transpose(grid),1E-7))
             rm = min(np.linalg.norm(min_grid,axis=0))
             pf = 4/3.*np.pi*rm**2/np.dot(min_grid[0],np.cross(min_grid[1],min_grid[2]))
-            grid={"grid_vecs":min_grid,"r_min":rm,"packing_frac":pg}
+            grid={"grid_vecs":min_grid,"r_min":rm,"packing_frac":pf}
             grids.append(grid)
 
     return grids
+